api/requirements.txt
fastapi==0.104.1
weaviate-client==4.4.0
python-dotenv==1.0.0
uvicorn==0.24.0
requests==2.31.0
daytona-sdk>=0.1.0
opik>=1.8.0